Fix gpu implementation
- Remove random variables
- Metal library needs to come from .metallib file
  which is generated

+++ random_walks/random_walk.metal
@@ -1,8 +1,9 @@
 #include <metal_stdlib>
+#include "mlx/backend/metal/kernels/utils.h"
 using namespace metal;
 
 kernel void random_walk(
-    const device int32_t* rowptr [[buffer(0)]],
+    const device int64_t* rowptr [[buffer(0)]],
     const device int32_t* col [[buffer(1)]],
     const device int32_t* start [[buffer(2)]],
     const device float* rand [[buffer(3)]],
@@ -15,15 +16,15 @@ kernel void random_walk(
     n_out[tid * (walk_length + 1)] = n_cur;
     
     for (int l = 0; l < walk_length; l++) {
-        int64_t row_start = rowptr[n_cur];
-        int64_t row_end = rowptr[n_cur + 1];
-        int64_t e_cur;
+        int32_t row_start = rowptr[n_cur];
+        int32_t row_end = rowptr[n_cur + 1];
+        int32_t e_cur;
         
         if (row_end - row_start == 0) {
             e_cur = -1;
         } else {
             float r = rand[tid * walk_length + l];
-            int64_t idx = static_cast<int64_t>(r * (row_end - row_start));
+            int32_t idx = static_cast<int32_t>(r * (row_end - row_start));
             e_cur = row_start + idx;
             n_cur = col[e_cur];
         }
